# onebrain-web/src/lib/openapi.json
{
  "components": {
    "schemas": {
      "AccessCheckRequest": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "app_id": {
            "title": "App Id",
            "type": "string"
          },
          "purpose": {
            "title": "Purpose",
            "type": "string"
          },
          "space_id": {
            "title": "Space Id",
            "type": "string"
          }
        },
        "required": [
          "account_id",
          "app_id",
          "space_id",
          "purpose"
        ],
        "title": "AccessCheckRequest",
        "type": "object"
      },
      "AccessCheckResponse": {
        "properties": {
          "allowed": {
            "title": "Allowed",
            "type": "boolean"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "allowed",
          "reason"
        ],
        "title": "AccessCheckResponse",
        "type": "object"
      },
      "AccountCreate": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "kind": {
            "pattern": "^(person|organization|family|project)$",
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "name"
        ],
        "title": "AccountCreate",
        "type": "object"
      },
      "AccountOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner_user_id": {
            "default": "",
            "title": "Owner User Id",
            "type": "string"
          },
          "status": {
            "default": "active",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "name"
        ],
        "title": "AccountOut",
        "type": "object"
      },
      "AppInstallCreate": {
        "properties": {
          "allowed_purposes": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Purposes",
            "type": "array"
          },
          "app_id": {
            "pattern": "^(onebrain_core|assistant|communication|admin_console|workers)$",
            "title": "App Id",
            "type": "string"
          },
          "display_name": {
            "default": "",
            "maxLength": 200,
            "title": "Display Name",
            "type": "string"
          },
          "enabled_space_ids": {
            "items": {
              "type": "string"
            },
            "title": "Enabled Space Ids",
            "type": "array"
          },
          "id": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          }
        },
        "required": [
          "app_id"
        ],
        "title": "AppInstallCreate",
        "type": "object"
      },
      "AppInstallationOut": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "allowed_purposes": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Purposes",
            "type": "array"
          },
          "app_id": {
            "title": "App Id",
            "type": "string"
          },
          "display_name": {
            "default": "",
            "title": "Display Name",
            "type": "string"
          },
          "enabled_space_ids": {
            "items": {
              "type": "string"
            },
            "title": "Enabled Space Ids",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "default": "active",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "account_id",
          "app_id",
          "enabled_space_ids",
          "allowed_purposes"
        ],
        "title": "AppInstallationOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SpaceCreate": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "kind": {
            "pattern": "^(personal|business|customer_service|shared|family|project)$",
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "name"
        ],
        "title": "SpaceCreate",
        "type": "object"
      },
      "SpaceOut": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "default": "active",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "account_id",
          "kind",
          "name"
        ],
        "title": "SpaceOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "onebrain",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/platform/access/check": {
      "post": {
        "operationId": "check_access_api_platform_access_check_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "ob_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ob Session",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessCheckResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check Access",
        "tags": [
          "platform"
        ]
      }
    },
    "/api/platform/accounts": {
      "get": {
        "operationId": "list_accounts_api_platform_accounts_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "ob_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ob Session",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AccountOut"
                  },
                  "title": "Response List Accounts Api Platform Accounts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Accounts",
        "tags": [
          "platform"
        ]
      },
      "post": {
        "operationId": "create_account_api_platform_accounts_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "ob_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ob Session",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Account",
        "tags": [
          "platform"
        ]
      }
    },
    "/api/platform/accounts/{account_id}/apps": {
      "get": {
        "operationId": "list_apps_api_platform_accounts__account_id__apps_get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "ob_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ob Session",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AppInstallationOut"
                  },
                  "title": "Response List Apps Api Platform Accounts  Account Id  Apps Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Apps",
        "tags": [
          "platform"
        ]
      },
      "post": {
        "operationId": "install_app_api_platform_accounts__account_id__apps_post",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "ob_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ob Session",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppInstallCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppInstallationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Install App",
        "tags": [
          "platform"
        ]
      }
    },
    "/api/platform/accounts/{account_id}/audit": {
      "get": {
        "operationId": "list_audit_api_platform_accounts__account_id__audit_get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "ob_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ob Session",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditOut"
                  },
                  "title": "Response List Audit Api Platform Accounts  Account Id  Audit Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Audit",
        "tags": [
          "platform"
        ]
      }
    },
    "/api/platform/accounts/{account_id}/brand-theme": {
      "get": {
        "operationId": "get_brand_theme_api_platform_accounts__account_id__brand_theme_get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "app_id",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 80,
              "title": "App Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "ob_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ob Session",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandThemeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Brand Theme",
        "tags": [
          "platform"
        ]
      },
      "put": {
        "operationId": "upsert_brand_theme_api_platform_accounts__account_id__brand_theme_put",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "ob_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ob Session",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__routers__platform__BrandThemeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandThemeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert Brand Theme",
        "tags": [
          "platform"
        ]
      }
    },
    "/api/platform/accounts/{account_id}/brand-themes": {
      "get": {
        "operationId": "list_brand_themes_api_platform_accounts__account_id__brand_themes_get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "ob_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ob Session",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BrandThemeOut"
                  },
                  "title": "Response List Brand Themes Api Platform Accounts  Account Id  Brand Themes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Brand Themes",
        "tags": [
          "platform"
        ]
      }
    },
    "/api/platform/accounts/{account_id}/spaces": {
      "get": {
        "operationId": "list_spaces_api_platform_accounts__account_id__spaces_get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "ob_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ob Session",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SpaceOut"
                  },
                  "title": "Response List Spaces Api Platform Accounts  Account Id  Spaces Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Spaces",
        "tags": [
          "platform"
        ]
      },
      "post": {
        "operationId": "create_space_api_platform_accounts__account_id__spaces_post",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "ob_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ob Session",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpaceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Space",
        "tags": [
          "platform"
        ]
      }
    },
...
  }
}